1. Project Packaging Hierarchy Restructured ... packaged created according to module like user / workorder 2. Loggin Added .... "logback.xml" file added for loggin configurations 3. Workorder Managament for procurement process classes added 4. New table added "workflow" for workorder generation automation framework
