+++ pom.xml
@@ -8,7 +8,7 @@
 		<version>2.7.8</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.example</groupId>
+	<groupId>com</groupId>
 	<artifactId>kapas</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>kapas</name>
