--- src/main/java/com/example/kapas/KapasApplication.java
@@ -1,15 +1,0 @@
-package com.example.kapas;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@SpringBootApplication
-@EnableScheduling
-public class KapasApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(KapasApplication.class, args);
-	}
-
-}
