--- src/main/java/com/example/kapas/entity/User.java
@@ -1,69 +1,0 @@
-package com.example.kapas.entity;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import java.time.Instant;
-
-@Table(name = "user")
-@Entity
-@Getter
-@Setter
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id", nullable = false)
-    private Integer id;
-
-    @Column(name = "user_name", nullable = false)
-    private String userName;
-
-    @Column(name = "first_name", nullable = false)
-    private String firstName;
-
-    @Column(name = "last_name", nullable = false)
-    private String lastName;
-
-    @Column(name = "email")
-    private String email;
-
-    @Column(name = "mobile", nullable = false)
-    private String mobile;
-
-    @Column(name = "password", nullable = false)
-    private String password;
-
-    @Column(name = "is_active", nullable = false)
-    private String isActive;
-
-    @Column(name = "description")
-    private String description;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "role_id", nullable = false)
-    private Role role;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "created_by", nullable = false)
-    private User createdBy;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "modified_by")
-    private User modifiedBy;
-
-    @Column(name = "creation_time", nullable = false)
-    private Instant creationTime;
-
-    @Column(name = "modification_time")
-    private Instant modificationTime;
-
-}
