--- src/main/java/com/example/kapas/model/ApiResponseError.java
@@ -1,13 +1,0 @@
-package com.example.kapas.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class ApiResponseError<T> {
-    String code;
-    T message;
-}
