--- src/main/java/com/example/kapas/model/UserSession.java
@@ -1,19 +1,0 @@
-package com.example.kapas.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserSession {
-
-    private String id;
-    private LocalDateTime expiresAt;
-    private LocalDateTime createdAt;
-}
