--- src/main/java/com/example/kapas/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.example.kapas.repository;
-
-import com.example.kapas.entity.User;
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends CrudRepository<User, Integer> {
-
-    Optional<User> findUserByUserName(String userName);
-}
