--- src/main/java/com/example/kapas/util/Utility.java
@@ -1,14 +1,0 @@
-package com.example.kapas.util;
-
-import java.util.concurrent.ThreadLocalRandom;
-
-public class Utility {
-
-    private Utility() {
-
-    }
-
-    public static int generateRandomNumber(int min, int max) {
-        return ThreadLocalRandom.current().nextInt(min, max + 1);
-    }
-}
