+++ src/main/java/com/kapas/user/entity/Permission.java
@@ -1,0 +1,53 @@
+package com.kapas.user.entity;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToMany;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+import java.time.Instant;
+import java.util.ArrayList;
+import java.util.List;
+
+@Table(name = "permission")
+@Entity
+@Getter
+@Setter
+public class Permission {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id", nullable = false)
+    private Integer id;
+
+    @Column(name = "permission_name", nullable = false)
+    private String permissionName;
+
+    @Column(name = "description")
+    private String description;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "created_by", nullable = false)
+    private User createdBy;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "modified_by")
+    private User modifiedBy;
+
+    @Column(name = "creation_time", nullable = false)
+    private Instant creationTime;
+
+    @Column(name = "modification_time")
+    private Instant modificationTime;
+
+    @ManyToMany(mappedBy = "permissions")
+    private List<Role> posts = new ArrayList<>();
+
+}
