+++ src/main/java/com/kapas/user/model/ApiResponseError.java
@@ -1,0 +1,13 @@
+package com.kapas.user.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class ApiResponseError<T> {
+    String code;
+    T message;
+}
