+++ src/main/java/com/kapas/user/model/UserSession.java
@@ -1,0 +1,19 @@
+package com.kapas.user.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class UserSession {
+
+    private String id;
+    private LocalDateTime expiresAt;
+    private LocalDateTime createdAt;
+}
