+++ src/main/java/com/kapas/user/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.kapas.user.repository;
+
+import com.kapas.user.entity.User;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends CrudRepository<User, Integer> {
+
+    Optional<User> findUserByUserName(String userName);
+}
