+++ src/main/java/com/kapas/workorder/entity/Workflow.java
@@ -1,0 +1,51 @@
+package com.kapas.workorder.entity;
+
+
+import com.kapas.user.entity.User;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.io.Serializable;
+import java.sql.Timestamp;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "workflow")
+public class Workflow implements Serializable {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id", nullable = false)
+    private Integer id;
+
+    private String workflow_id;
+
+    private String name;
+
+    private Boolean is_active;
+
+    private Integer version;
+
+    private String description;
+
+    private String workflow_json;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "created_by", nullable = false)
+    private User created_by;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "modified_by", nullable = false)
+    private User modified_by;
+
+    private Timestamp creation_time;
+
+    private Timestamp modification_time;
+
+}
