+++ src/main/java/com/kapas/workorder/model/MetaDataField.java
@@ -1,0 +1,19 @@
+package com.kapas.workorder.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class MetaDataField {
+
+    private String key;
+
+    private String value;
+
+    private Boolean is_requierd;
+}
