+++ src/main/java/com/kapas/workorder/service/WorkorderService.java
@@ -1,0 +1,13 @@
+package com.kapas.workorder.service;
+
+import com.kapas.workorder.entity.Workorder;
+import org.springframework.stereotype.Service;
+
+@Service
+public class WorkorderService {
+
+    private Workorder createWorkorder() {
+        return null;
+    }
+
+}
