--- src/test/java/com/example/kapas/KapasApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.kapas;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class KapasApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
