+++ src/test/java/com/kapas/KapasApplicationTests.java
@@ -1,0 +1,13 @@
+package com.kapas;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class KapasApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
